Create a separate chunk for watch mode via a dynamic import
