+++ src/browser-entry.ts
@@ -1,2 +1,2 @@
-export { default as rollup } from './rollup/index';
+export { default as rollup } from './rollup/rollup';
 export { version as VERSION } from 'package.json';
